backend/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
python-dotenv==1.0.1
google-generativeai==0.8.4
faiss-cpu==1.9.0.post1
numpy==2.2.3
pydantic==2.10.5
python-jose[cryptography]==3.4.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
slowapi==0.1.9
sqlalchemy==2.0.36
psycopg2-binary==2.9.9
pytest==8.0.2
httpx==0.27.0
langchain==0.1.13
langchain-community==0.0.29
pypdf==4.1.01
